+++ main.tf
@@ -3,6 +3,7 @@ resource "azurerm_application_gateway" "main" {
   resource_group_name = var.resource_group_name
   location            = var.location
   tags                = var.tags
+  zones               = var.zones
   firewall_policy_id  = var.firewall_policy_id
 
   sku {